Added FSRS scheduler and basic card rescheduling functionality; Add basic DB functions
TODO: Sync the database with creating new cards and resechduling existing cards. The pull button
should get the scheduled card from the database (or we have a structure that pulls all the cards for
the day so we aren't hitting the db over and over), and the create button should create a new card to
be inserted into the database.